--- mcp_tools/binary_analysis/binary_analysis_and_reverse_engineering.py
@@ -1,107 +1,0 @@
-# mcp_tools/binary_analysis_and_reverse_engineering.py
-
-from typing import Dict, Any
-
-def register_binary_analysis_and_reverse_engineering_tools(mcp, hexstrike_client, logger):
-
-    @mcp.tool()
-    def checksec_analyze(binary: str) -> Dict[str, Any]:
-        """
-        Check security features of a binary with enhanced logging.
-
-        Args:
-            binary: Path to the binary file
-
-        Returns:
-            Security features analysis results
-        """
-        data = {
-            "binary": binary
-        }
-        logger.info(f"🔧 Starting Checksec analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/checksec", data)
-        if result.get("success"):
-            logger.info(f"✅ Checksec analysis completed for {binary}")
-        else:
-            logger.error(f"❌ Checksec analysis failed for {binary}")
-        return result
-
-    @mcp.tool()
-    def xxd_hexdump(file_path: str, offset: str = "0", length: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Create a hex dump of a file using xxd with enhanced logging.
-
-        Args:
-            file_path: Path to the file
-            offset: Offset to start reading from
-            length: Number of bytes to read
-            additional_args: Additional xxd arguments
-
-        Returns:
-            Hex dump results
-        """
-        data = {
-            "file_path": file_path,
-            "offset": offset,
-            "length": length,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting XXD hex dump: {file_path}")
-        result = hexstrike_client.safe_post("api/tools/xxd", data)
-        if result.get("success"):
-            logger.info(f"✅ XXD hex dump completed for {file_path}")
-        else:
-            logger.error(f"❌ XXD hex dump failed for {file_path}")
-        return result
-
-    @mcp.tool()
-    def strings_extract(file_path: str, min_len: int = 4, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Extract strings from a binary file with enhanced logging.
-
-        Args:
-            file_path: Path to the file
-            min_len: Minimum string length
-            additional_args: Additional strings arguments
-
-        Returns:
-            String extraction results
-        """
-        data = {
-            "file_path": file_path,
-            "min_len": min_len,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Strings extraction: {file_path}")
-        result = hexstrike_client.safe_post("api/tools/strings", data)
-        if result.get("success"):
-            logger.info(f"✅ Strings extraction completed for {file_path}")
-        else:
-            logger.error(f"❌ Strings extraction failed for {file_path}")
-        return result
-
-    @mcp.tool()
-    def objdump_analyze(binary: str, disassemble: bool = True, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Analyze a binary using objdump with enhanced logging.
-
-        Args:
-            binary: Path to the binary file
-            disassemble: Whether to disassemble the binary
-            additional_args: Additional objdump arguments
-
-        Returns:
-            Binary analysis results
-        """
-        data = {
-            "binary": binary,
-            "disassemble": disassemble,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔧 Starting Objdump analysis: {binary}")
-        result = hexstrike_client.safe_post("api/tools/objdump", data)
-        if result.get("success"):
-            logger.info(f"✅ Objdump analysis completed for {binary}")
-        else:
-            logger.error(f"❌ Objdump analysis failed for {binary}")
-        return result
